requirements.txt
flask
google-genai
PyPDF2
python-dotenv